requirements.txt
wheel
Celery
supervisor
redis
Flask
Flask-WTF
Flask-SQLAlchemy
Flask-Login
Flask-Session
Flask-Migrate
Flask-Mail
python-dotenv
email-validator
PyMySQL
flower